+++ android/app/src/main/kotlin/com/example/demo_using_pkg/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.demo_using_pkg
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}
