--- android/app/src/main/kotlin/com/example/pkg_text_field/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.pkg_text_field
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}
