--- lib/const/constat.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Const {
-
-  static Type ifelse<Type>(
-      bool condition, {
-        required Type valid,
-        required Type invalid,
-      }) {
-    return condition ? valid : invalid;
-  }
-}
