--- lib/info.text
@@ -1,6 +1,0 @@
-
-TODO : BIT BUCKET LINK : https://bitbucket.org/flutter-demo-parth/hardware_asset_management/src/master/
-latest code available in github
-
-flutter version with fvm
-fvm use 3.10.5
