+++ lib/main.dart
@@ -1,18 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:pkg_text_field/UI/suggest_textfield/database/databasehelper.dart';
-import 'package:pkg_text_field/UI/view/screen/suggetion_api_view.dart';
-import 'package:sqflite/sqflite.dart';
-import 'package:flutter/services.dart';
+import 'package:suggestion_textfield/suggestion_textfield.dart';
 
 void main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-  DatabaseHelper helper = DatabaseHelper();
-  Database? db;
-  db = await helper.getDB();
-  debugPrint('db : $db');
-  SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual,
-      overlays: [SystemUiOverlay.bottom, SystemUiOverlay.top]);
-
   runApp(const MyApp());
 }
 
@@ -25,10 +14,67 @@ class MyApp extends StatelessWidget {
     return const MaterialApp(
       title: 'Flutter Demo',
       debugShowCheckedModeBanner: false,
-      home: SuggetionApiPage(),
+      home: HomePage(),
     );
   }
 }
 
+class HomePage extends StatefulWidget {
+  const HomePage({Key? key}) : super(key: key);
+
+  @override
+  State<HomePage> createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+
+  FocusNode focusNode = FocusNode();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text(
+          'Suggestion Text field',
+        ),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 16,vertical: 10),
+        child: ListView(
+          children: [
+            SuggestionTextField(
+              focusNode: focusNode,
+              // controller: cDictionary,
+              // onChange: (onChangeVal) {
+              //   debugPrint('onChangeVal : $onChangeVal');
+              // },
+              onTap: () {
+                debugPrint(' ===========> onTap <=========== ');
+              },
+              onChangeValue: (onchangeValCustom) {
+                debugPrint('onchangeValCustom : $onchangeValCustom');
+              },
+              maxElementsToDisplay: 50,
+              showSuggestionList: true,
+            ),
+            ListView.builder(
+              itemCount: 10,
+              shrinkWrap: true,
+              physics: const NeverScrollableScrollPhysics(),
+              itemBuilder: (context, index) {
+                return Column(
+                  children: const [
+                    Text("Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.\n\n"),
+                    SizedBox(height: 10),
+                  ],
+                );
+              },
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
