+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: pkg_text_field
+name: demo_using_pkg
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
@@ -34,8 +34,10 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  sqflite: ^2.0.2
-  path_provider: ^2.0.9
+  suggestion_textfield:
+    git: 
+      url: https://github.com/parthunagar/SuggestionTextField.git
+    # path: /Users/knoxpo/Documents/parthunagar/project/package/suggestion_textfield_git
 
 dev_dependencies:
   flutter_test:
@@ -60,8 +62,6 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  assets:
-    - assets/database/
   # assets:
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
